pyproject.toml
[project]
name = "doc-parser-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
